+++ mcs/class/Facades/System.Diagnostics.Tracing/System.Diagnostics.Tracing.dll.sources
@@ -1,3 +1,2 @@
 TypeForwarders.cs
 AssemblyInfo.cs
-EventCounter.cs
