+++ mcs/class/Facades/System.Net.Sockets/System.Net.Sockets.dll.sources
@@ -1,5 +1,2 @@
 TypeForwarders.cs
 AssemblyInfo.cs
-SocketReceiveFromResult.cs
-SocketReceiveMessageFromResult.cs
-SocketTaskExtensions.cs
