+++ mcs/class/System/System.Net.Sockets/SocketTaskExtensions.cs
@@ -2,6 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+#if NETSTANDARD
+
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -248,3 +250,5 @@ namespace System.Net.Sockets
         }
     }
 }
+
+#endif
